refactor: response 파일 패키지 이동 (#175)
